+++ ssl/ssl_local.h
@@ -43,6 +43,8 @@
 #  define OPENSSL_EXTERN OPENSSL_EXPORT
 # endif
 
+DEFINE_STACK_OF(EVP_PKEY)
+
 # define c2l(c,l)        (l = ((unsigned long)(*((c)++)))     , \
                          l|=(((unsigned long)(*((c)++)))<< 8), \
                          l|=(((unsigned long)(*((c)++)))<<16), \
@@ -591,6 +593,9 @@ struct ssl_session_st {
      * to disable session caching and tickets.
      */
     int not_resumable;
+# ifndef OPENSSL_NO_RPK
+    EVP_PKEY *peer_rpk;
+# endif
     /* This is the cert and type for the other end. */
     X509 *peer;
     /* Certificate chain peer sent. */
@@ -765,6 +770,8 @@ typedef enum tlsext_index_en {
     TLSEXT_IDX_extended_master_secret,
     TLSEXT_IDX_signature_algorithms_cert,
     TLSEXT_IDX_post_handshake_auth,
+    TLSEXT_IDX_client_cert_type,
+    TLSEXT_IDX_server_cert_type,
     TLSEXT_IDX_signature_algorithms,
     TLSEXT_IDX_supported_versions,
     TLSEXT_IDX_psk_kex_modes,
@@ -1563,6 +1570,10 @@ struct ssl_connection_st {
     /* Verified chain of peer */
     STACK_OF(X509) *verified_chain;
     long verify_result;
+# ifndef OPENSSL_NO_RPK
+    /* The configured raw public keys we will accept from the peer */
+    STACK_OF(EVP_PKEY) *peer_rpks;
+# endif
     /*
      * What we put in certificate_authorities extension for TLS 1.3
      * (ClientHello and CertificateRequest) or just client cert requests for
@@ -1701,6 +1712,12 @@ struct ssl_connection_st {
          * selected.
          */
         int tick_identity;
+#ifndef OPENSSL_NO_RPK
+        uint8_t client_cert_type;
+        uint8_t client_cert_type_ctos;
+        uint8_t server_cert_type;
+        uint8_t server_cert_type_ctos;
+#endif
     } ext;
 
     /*
@@ -2426,11 +2443,33 @@ struct openssl_ssl_test_functions {
 
 const char *ssl_protocol_to_string(int version);
 
+#ifndef OPENSSL_NO_RPK
+static ossl_inline int tls12_rpk_and_privkey(const SSL_CONNECTION *sc, int idx)
+{
+    /*
+     * This is to check for special cases when using RPK with just
+     * a private key, and NO CERTIFICATE
+     */
+    return ((sc->server && sc->ext.server_cert_type == TLSEXT_cert_type_rpk)
+            || (!sc->server && sc->ext.client_cert_type == TLSEXT_cert_type_rpk))
+        && sc->cert->pkeys[idx].privatekey != NULL
+        && sc->cert->pkeys[idx].x509 == NULL;
+}
+
+#endif
+
 /* Returns true if certificate and private key for 'idx' are present */
 static ossl_inline int ssl_has_cert(const SSL_CONNECTION *s, int idx)
 {
     if (idx < 0 || idx >= SSL_PKEY_NUM)
         return 0;
+#ifndef OPENSSL_NO_RPK
+    /* If RPK is enabled for this SSL... only require private key */
+    if ((s->server && (s->options & SSL_OP_RPK_SERVER))
+            || (!s->server && (s->options & SSL_OP_RPK_CLIENT))) {
+        return s->cert->pkeys[idx].privatekey != NULL;
+    }
+#endif
     return s->cert->pkeys[idx].x509 != NULL
         && s->cert->pkeys[idx].privatekey != NULL;
 }