+++ CHANGES.md
@@ -24,6 +24,10 @@ OpenSSL 3.1
 
 ### Changes between 3.0 and 3.1 [xx XXX xxxx]
 
+ * Add Raw Public Key (RFC7250) support.
+
+   *Todd Short*
+
  * Major refactor of the libssl record layer
 
    *Matt Caswell*